requirements.txt
discord.py
yt-dlp
Flask
PyNaCl